refactor: extract IssueExtractionResult.swift from IssueExportTargets.swift
Splits the `IssueExtractionResult` value (extraction-pipeline output)
out of IssueExportTargets.swift so that file holds only export-related
target types.

Byte-identical move of lines 66-88. Only imports Foundation.

IssueExportTargets.swift: 124 → 100 lines. Tests: 667.
