+++ Sources/BugNarrator/Models/IssueExportTargets.swift
@@ -62,30 +62,6 @@ struct JiraIssueExportTarget: Codable, Equatable {
         return "\(projectKey) - \(projectName)"
     }
 }
-
-struct IssueExtractionResult: Codable, Equatable {
-    let generatedAt: Date
-    var summary: String
-    var guidanceNote: String
-    var issues: [ExtractedIssue]
-
-    init(
-        generatedAt: Date = Date(),
-        summary: String,
-        guidanceNote: String = "Extracted issues are draft suggestions and should be reviewed before export.",
-        issues: [ExtractedIssue]
-    ) {
-        self.generatedAt = generatedAt
-        self.summary = summary
-        self.guidanceNote = guidanceNote
-        self.issues = issues
-    }
-
-    var selectedIssues: [ExtractedIssue] {
-        issues.filter(\.isSelectedForExport)
-    }
-}
-
 enum ExportDestination: String, Codable, CaseIterable, Identifiable {
     case github = "GitHub"
     case jira = "Jira"
