+++ Sources/BugNarrator/Models/IssueExtractionResult.swift
@@ -1,0 +1,25 @@
+import Foundation
+
+struct IssueExtractionResult: Codable, Equatable {
+    let generatedAt: Date
+    var summary: String
+    var guidanceNote: String
+    var issues: [ExtractedIssue]
+
+    init(
+        generatedAt: Date = Date(),
+        summary: String,
+        guidanceNote: String = "Extracted issues are draft suggestions and should be reviewed before export.",
+        issues: [ExtractedIssue]
+    ) {
+        self.generatedAt = generatedAt
+        self.summary = summary
+        self.guidanceNote = guidanceNote
+        self.issues = issues
+    }
+
+    var selectedIssues: [ExtractedIssue] {
+        issues.filter(\.isSelectedForExport)
+    }
+}
+
